requirements.txt
numpy==1.26.4
pillow
tqdm
albumentations==1.3.1
albucore==0.0.11
scikit-image
opencv-python
segmentation-models-pytorch==0.3.3
pretrainedmodels
six
matplotlib
imageio
